+++ lib/rodauth/rails/railtie.rb
@@ -1,6 +1,5 @@
 require "rodauth/rails/middleware"
 require "rodauth/rails/controller_methods"
-require "rodauth/rails/log_subscriber"
 
 require "rails"
 
@@ -15,10 +14,6 @@ module Rodauth
         ActiveSupport.on_load(:action_controller) do
           include Rodauth::Rails::ControllerMethods
         end
-      end
-
-      initializer "rodauth.log_subscriber" do
-        Rodauth::Rails::LogSubscriber.attach_to :rodauth
       end
 
       initializer "rodauth.test" do