+++ test/integration/instrumentation_test.rb
@@ -1,13 +1,12 @@
 require "test_helper"
 
-class LoggingTest < IntegrationTest
+class InstrumentationTest < IntegrationTest
   test "logs processing and completion of a request" do
     logged = capture_log do
       visit "/login"
     end
 
-    assert_match /Started GET "\/login" for 127\.0\.0\.1/, logged
-    assert_match /Processing by RodauthApp as HTML/, logged
+    assert_match /Processing by RodauthApp#call as HTML/, logged
     refute_match /Parameters/, logged
     if ::Rails.gem_version >= Gem::Version.new("6.0")
       assert_match /Completed 200 OK in \d+ms \(ActiveRecord: \d+\.\d+ms | Allocations: \d+\)/, logged
@@ -23,8 +22,8 @@ class LoggingTest < IntegrationTest
         { "CONTENT_TYPE" => "application/json", "HTTP_ACCEPT" => "application/json" }
     end
 
-    assert_match /Processing by RodauthApp as JSON/, logged
-    assert_match /\s{2}Parameters: {"login"=>"user@example\.com", "password"=>"secret"}/, logged
+    assert_match /Processing by RodauthApp#call as JSON/, logged
+    assert_match /Parameters: {"login"=>"user@example\.com", "password"=>"secret"}/, logged
     assert_match /Completed 401 Unauthorized/, logged
   end
 
@@ -33,8 +32,7 @@ class LoggingTest < IntegrationTest
       visit "/login?early_return=true"
     end
 
-    assert_match /Started GET "\/login\?early_return=true" for 127\.0\.0\.1/, logged
-    assert_match /Processing by RodauthApp as HTML/, logged
+    assert_match /Processing by RodauthApp#call as HTML/, logged
     assert_match /Completed 201 Created in \d+ms/, logged
   end
 
@@ -50,10 +48,10 @@ class LoggingTest < IntegrationTest
 
   def capture_log
     io = StringIO.new
-    original_logger = Rails.logger
-    Rails.logger = Logger.new(io)
+    original_logger = ActionController::Base.logger
+    ActionController::Base.logger = Logger.new(io)
     yield
-    Rails.logger = original_logger
+    ActionController::Base.logger = original_logger
     io.string
   end
 end