+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Instrument Rodauth requests on `action_controller` namespace (@janko)
+
 * Update templates for Boostratp 5 compatibility (@janko)
 
 * Log request parameters for Rodauth requests (@janko)