+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 0.12.0 (2021-05-15)
 
 * Include total view render time in logs for Rodauth requests (@janko)